Simplify language around `pkgPathFor` helper.
This clarifies the language around usage of the `pkgPathFor` helper by
explicitly indicating the `pkg_deps` requirement and providing a
practical example which can be easily followed and understood.
Additionally, this adds some information about what will happen when a
package is not in `pkg_deps` so that the user can identify the
issue and resolve the problem quickly.

Signed-off-by: Ryan Hass <rhass@users.noreply.github.com>