Fix `/services/{svc}/{group}/config` endpoint in http-gateway
The config endpoint was recently broken from refactoring how we
serialize information in the supervisor in a previous commit. We
no longer write out a `config.toml` with what was our old rendering
context.

This adds functionality to the endpoint once again but this time
only the actual configuration of the service is returned

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>