Use protobuf to represent sys information in swim.Service proto
Prior to this change we were serializing a SysInfo struct into
TOML and then passing it across the wire as bytes. This isn't as
resilient to protocol changes as using a protobuf; it's less efficient
and it makes it more difficult to version our protocol in a way to
allow multiple supervisor versions to gossip to each other.

This is a breaking change. A new version of the supervisor will need
to be cut and we must increment the minor version. Previous versions
of the supervisor will not be able to communicate with ones built
after this commit. Fortunately this should stop us from introducing
further breaking changes (with this message at least).

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>