Relax authorization checks on Builder API read-only endpoints
For the initial rollout of the Builder service, we are only enabling
it for the `core` origin; no other origins will be using it to build
artifacts. Since `core` origin is a community resource, anybody should
be able to see project and build job information (e.g., build logs!).

As such, we can relax the authorization checks on these read-only
endpoints. You just need to be logged in with your Github credentials,
but you don't need to have further permissions beyond that.

Signed-off-by: Christopher Maier <cmaier@chef.io>