Fix log-streaming performance
Currently the build log is being rendered with Angular’s default change-detection behavior, which is causing it to be rendered much more often than it needs to be. This change adds an RxJS BehaviorSubject to allow subscribers to be notified when log segments are received, and updates the log component to append fragments to the log element rather than replace its content completely.

Signed-off-by: Christian Nunciato <cnunciato@chef.io>