[studio] Refactor the auto-running Supervisor experience.
This change refactors some of the implementation which runs a Supervisor
when entering an interactive Studio, but also adds a mechanism by which
you can terminate the Supervisor and customize it's run arguments via an
environment variable.

New Environment Variable
------------------------

A new environment can be passed to `hab studio enter` called
`HAB_STUDIO_SUP` which does 2 things: provide a way to not run a
Supervisor or to customize how the Supervisor runs by passing additional
command line arguments such as `--peer`, etc.

When set to `false`, `no`, or `0`, a Supervisor will not run
automatically in an interactive Studio, allowing a user to
export-and-forget if that is their preference.

Alternatively, when non-empty (and not set to a value above), it is
assumed that the value contains extra command line arguments that will
be passed to `hab sup run`. For example, to interactive enter a Studio
with a Supervisor running, joined to a ring you could run:

    env HAB_STUDIO_SUP='--peer 172.17.0.2' hab studio enter

New In-Studio Helper Functions
------------------------------

Three new functions are set inside an interactive Studio session:

* `sup-log`  - Tails the Supervisor's output (this replaces `slog`)
* `sup-term` - Terminates the Supervisor running in the background
* `sup-run`  - Runs a Supervisor in the background. You can pass
               additional arguments such as `--peer`, etc. This is what
               is called to automatically run the Supervisor in
               the background.

New Shortcut Aliases
--------------------

If you desire less typing (I sure do), there are also three new shell
aliases:

* `alias sl='sup-log'`
* `alias st='sup-term'`
* `alias sr='sup-run'`

Yes. You've just taught yourself to not make the `sl` typo and now
we're letting you type it again. Life keeps throwin' curveballs.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>