CensusMember is alive if it is ourself
When we construct a new CensusMember we always use the default trait.
The default trait will set the `alive` field to false. When a service
rumor is exchanged into a CensusMember we would previously not check
to see if the rumor originated from ourself.

With this change any CensusMember that we inject will first check
to see if it originated from our member_id. If it has, then we will
mark the CensusMember as alive.

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>