Create bucket for job logs
This creates an S3 bucket to store job logs in, along with an IAM user
and a bucket policy that only allows that user to put and get objects
from the bucket.

To configure the job server to use the bucket, we will need to create
a key and secret for this IAM user out-of-band, but we may be able to
use instance roles in the near-ish future.

Signed-off-by: Christopher Maier <cmaier@chef.io>