Add HAB_UPDATE_STRATEGY_FREQUENCY_MS to studio propagated vars
This is needed to tell the supervisor how often to check for updates,
which, in a studio, you often want to do right away after you've built a
package.

This is not available as a command line option to `hab sup run`, so it
cannot be set via the `HAB_STUDIO_SUP` options. Therefore, we make it so
it can be set outside the studio.

Signed-off-by: Nathan L Smith <smith@chef.io>