[sup] Introduce optional application environment qualifier on ServiceGroup.
This change adds a new optional qualifier to a Service Group string,
called an Application Environment. The printed representation of a
Service Group with Application Environment delimited with a hash
character and the application name and environment name are delimited
with a period character. For example, a Service Group containing an
application name "twitter" and environment name of "production" would
look like:

    twitter.production#redis.cache

The components of the Service Group are as follows:

        twitter   . production  #  redis  . cache
    |-application-|-environment-|-service-|-group-|

To load a service, 2 new options are added which are used together
(`--application/-a` and `--environment/-e`). For example:

    hab svc load core/redis \
      --group cache \
      --application twitter \
      --environment production

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>