Implement "hab job start" command.
This command lets the user start a job manually by specifying a package
identifier to be built, as well as whether they want to just build that package
or that package and all of its reverse dependencies as well. If they choose to
do the entire group, they will be shown what will be built first and prompted
to confirm that's what they want.

This change also introduces the builder-api-client crate for communicating with
builder-api. This is similar to the builder-depot-client crate but is specific
to the URL and routes available in builder-api.

Signed-off-by: Josh Black <raskchanky@gmail.com>