Fix intermittent spec-watcher test failures
We keep getting periodic test failures that seem to be traced back to
this timeout. I suspect that 100ms is just too short. I've bumped the
overall limit up to 1 second, but since the test actually checks for
new events every millisecond, this shouldn't make tests run any
longer; it'll just give us a buffer on testing systems that don't
quite as fast as we'd like.

Signed-off-by: Christopher Maier <cmaier@chef.io>