Add some smarts around picking random ports for test supervisors
Here, we add a mutex-protected hash set of all the ports that our
test supervisors are using. We also check to see that the port we pick
isn't already bound by some other process on the system. We've even
got retry logic, too.

All this allows us to be reasonably sure that the ports we choose
aren't going to conflict with anything else, and that our tests won't
spuriously fail.

Signed-off-by: Christopher Maier <cmaier@chef.io>