`hab svc status` returns correct answer regardless of executing user
Checking if the Supervisor was alive would be incorrectly reported if
a user without the same privileges as the process which started the
Supervisor ran the `hab svc status` command. This command isn't
privileged so a few minor tweaks to the `is_alive()` function in
the core crate should do the trick.

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>