[studio] Kill Launcher process when shutting down Studio.
This change follows the introduction of the Launcher in the 0.26.0
release. Prior to this commit the Studio cleanup logic would attempt to
kill the Supevisor but with the introduction of the Launcher, this would
leave a `hab-launch` process behind and running in Linux Studios. With
this change, the process ID of the Launcher is recorded in a file and is
checked on Studio shutdown for which process to kill. As the Launcher is
now the parent process of the Supervisor, the full chain of processes
should terminate as expected.

When making this change, I made a couple of refactoring around the
`cleanup_studio()` function including extracting the 3 cleanup tasks
into their own functions and fixed on bug in `studio rm` where a local
variable `$v` was missing.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>