Install initial habitat packages from bootstrap bundle
Rebuilding a Habitat Builder server instance in production is
currently harder than it needs to be, mainly because we've been using
Builder to bootstrap! If you need to replace a server in production,
chances are that the Builder isn't gonna be there for you. What to do?

Well, we can have a bundle of all the artifacts we need to get an
entire Builder infrastructure that's *not* coming from Builder, and
install from *that*. The solution has two parts. First, the
`create_bootstrap_bundle.sh` script is used to download all the
packages from a (currently functional!) Builder instance, `tar` them
up, and put them up into S3. This script is something that we can run
periodically (possibly with each release of the supervisor). All
bootstrap bundles created this way stay in S3, so it's theoretically
possible to stand up a Builder infrastructure from code at different
points in time, should we ever need / want to do such a
thing. Crucially, the script generates a manifest file for each
bundle, and also uploads this to S3, with a copy saved to the file
`LATEST`. All we need to do is grab the `LATEST` file, and it'll tell
us which tarball we should grab to bootstrap from.

The second part of the solution is to tweak our Terraform setup to
install packages from this tarball; that's largely the job of the
`install_base_packages.sh` script. It grabs the appropriate tarball on
every server instance and installs the desired packages into the local
`/hab/cache/artifacts` directory. With those packages in place, we can
run `hab svc load` for all our services without having to talk to
another Builder instance at all.

The previous `bootstrap.sh` script was removed, and the user creation
tasks it did were moved into the `filesystem.sh` script, which was
renamed to `foundation.sh` to reflect its more general purpose. This
was done to consolidate scripts.

Fixes #2732

Signed-off-by: Christopher Maier <cmaier@chef.io>