Fix panic when sending rumors of departed members
This PR is largely a change to the signature of
`member::membership_for`. We previously asserted that there was no
condition where a member could not be present in the member list
which we had never heard about. With the introduction of departed
members, we now have this case to deal with so `membership_for` and
any calling functions will now handle an Option case

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>