Release Launcher's LOCK on Supervisor restart
Added an environment variable to signal the Supervisor to release the
Launcher's process lock before attempting to obtain a new one. This
is useful for when a Supervisor shuts down gracefully and is restarted
by the Launcher. Prior to this change the Supervisor would fail to
start and exit with an `ERR_NO_RETRY_EXCODE`

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>