[studio] Use system artifact cache dir when running studio as root.
This change was motivated by a bug fix to the artifact caching feature
introduced in #2737 but upon closer inspection, a much better
refactoring presented itself.

The key bit of insight was to remember that on Linux systems where
`sudo` will be called, we can set up the appropriate environment
variables *before* we become the root user (under the hood we call
`exec` with sudo to re-run the program with higher permissions). Now
that all the home directory-related logic is run under the non-root
user, there are far fewer edge conditions and we no longer need the
logic which attempt to determine that original calling user via the
`SUDO_USER` environment variable.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>