Extract rumor "heat" tracking to separate module
This extracts the old `RumorList` struct, renames it to `RumorHeat`,
and simplifies its API to better reflect what its doing in such a way
that calling code reads more clearly.

Calling it a "rumor list" was misleading because a) it's not a list!
and b) it gave no indication that it was actually being used to track
the "heat" of a rumor. The API the struct exposed was also not as
descriptive as it could be, it returned more data than necessary, and
it also had more functions than we truly needed!

Particular changes to note:

* the `take` function was removed; it was only used in tests, and the
  tests were rewritten to not need it.
* the `take_by_kind` function was removed; it was only called in one
  place, so the implementation was basically just moved to that one
  call site.
* `insert` was renamed `start_hot_rumor` to better reflect what it
  actually did (in particular, calling it a second time didn't just
  "insert" a rumor, it effectively reset the "heat" information if the
  rumor was already being tracked).
* The `rumors` function was renamed to `currently_hot_rumors` to
  better reflect what it was doing and make calling code read
  better. Additionally, it now returns a `Vec<RumorKey>` instead of a
  `RumorVec`, which was a vector of tuples that paired a `RumorKey`
  with the current "heat" of that rumor. It turns out, this additional
  "heat" information was never used at call sites; it was really a
  purely internal detail. All call sites were correspondingly changed.
* `update_heat` was renamed to `cool_rumors` to better express what it
  was doing.
* Tests were revamped.

Signed-off-by: Christopher Maier <cmaier@chef.io>