Don't upload to "stable" channel by default
When packages are uploaded, they *always* get added to the `unstable`
channel. However, by passing the `--channel` option to `hab pkg
upload`, users may specify an additional release channel to put the
package into.

Previously, we erroneously defaulted the value of this option to
`stable`, meaning that unless the user specified something else, every
package would be uploaded to both the `unstable` *and* `stable`
channels.

The fix turned out to be pretty easy. We were already passing around
an `Option` type for the channel name; now, that is simply a `None` if
the user doesn't supply a value. Docstrings, comments, and variable
names were changed to reflect this new state of affairs and make it
more clear what the desired behavior is.

Signed-off-by: Christopher Maier <cmaier@chef.io>