Modified hab-plan-build to search subdirectory tree for plan.{sh,ps1}
The way hab-plan-build currently searches for the plan file is to check the root of the context and a folder named 'habitat' in the root of the context. While this is reasonable, it is also inflexible. This commit adds a dependency to saffronsnail/find-file in order to easily search for all files named 'plan.sh' (or plan.ps1, in the case of Windows) and prompt the user to select a plan file to use.

Signed-off-by: Bryan Ferris <primummoven@gmail.com>