[sup/functional] Print sup output in tests if `--nocapture` is set.
This change helps to better debug when functional supervsior tests are
failing. Existing behavior is preserved when running the tests; that is
to not output supervisor messages. However, if a developer wishes to see
more test output using the standard `--nocapture` flag, then the
supervisor output will be emitted.

To see the difference you can run:

    cd components/sup
    cargo test -- --nocapture

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>