[sup/functional] Support running functionals as a non-root user.
For the current functional tests which exercise hook file creation, the
service user running the service is not critical and not the subject
under test. Therefore, when preparing a fixture package, the following
package metafiles will be created *unless* one is present in the fixture
directory:

* `SVC_USER`: defaults to the current user running the tests
* `SVC_GROUP`: defaults to the current user's primary group running the
  tests

This approach will pass the service user/group assertions that the
Supervisor makes when creating the service directories without changing
production/runtime behavior.

As a result, and with the inclusion of #3022, all doc, unit, and
functional tests for the Supervisor can be run with a non-root user.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>