[sup] Fix hook table unit test which relied on lazy static value.
This change fixes a unit test that would not consistently pass or fail
when run as a non-root user. Prior to this change, the
`HookTable::load()` function computed its hook path using the
`sup::fs::svc_hooks_path()` which uses a `lazy_static` constant to
determine the root filesystem path for all future calls.

Rust unit test suites run many tests concurrently in threads so that
ordering is nondeterministic. Since all threads share the same
constants, the first test to fire code which calls-and-sets the
`core::fs::FS_ROOT_PATH` constant will "win" and all other threads will
use this value. The test in question attempted to set the filesystem
root to a temporary path and so would sometimes "win" and sometimes
"lose".

This change changes the signature of `HookTable::load()` to take the
path explicitly, leading to easer test setup and a more explicit
dependency injection for the production Supervisor code.

Also, this test shouldn't fail randomly any more. Not bad, eh?

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>