Plumb integrations through to the worker.
The way this works is the jobsrv retrieves all integrations for an origin when
it pulls a job off the queue. It decrypts everything and adds the unencrypted
integration information directly to the job. It then passes the job off to the
worker as normal, and the worker has that information to do whatever it needs.
This way, the unencrypted information is never persisted anywhere.

Signed-off-by: Josh Black <raskchanky@gmail.com>