Add FullyQualifiedPackageIdentRequired error
Previously, the internal `read_deps` function would return an
`InvalidPackageIdent` error if a dependency in a package metadata
file (e.g. `DEPS`, `TDEPS`) was not fully qualified. This is the wrong
error in this case, because we will have successfully parsed the
identifier string; _that_ is the scenario the `InvalidPackageIdent`
error covers. It also yields a confusing error message in that
case (basically "you gave us 'foo/bar'; you should use something like
'foo/bar' instead").

Now, we have a distinct error to capture the violated requirement of
having fully-qualified identifiers.

Signed-off-by: Christopher Maier <cmaier@chef.io>