[hab] Install minimum version of exporters when used.
This change updates the install-on-use behavior for all packages which
provide package export functionality (i.e. the implementations for `hab
pkg export *`). The behavior now is the same as the behavior for `hab
sup`, `hab run`, `hab svc *`, `hab config apply`, `hab pkg build`, and
`hab studio *` subcommands, where the same x.y.z version installed if
missing.

Prior to this change, upgrading the version of `core/hab` would not
upgrade the version of `core/hab-pkg-dockerize` (or other export
packages).

Note that in this change, 2 options were removed that affected the `hab
pkg export` subcommand, namely `--hab-url` and `hab-channel`. These
options were only used to install the `core/hab-pkg-dockerize` (and
similar) package itself which should not be required. Most importantly,
those options did not affect or influence which version of `core/hab`,
`core/hab-launcher`, and `core/hab-sup` were exported various export
types. A future change will introduce a re-written `hab-pkg-dockerize`
which fully supports targetting a different Depot URL and channel for
"base packages" such as the Supervisor and the Launcher.

Closes #3093

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>