Consolidate and refactor package installation logic
We have several places in the codebase that need to install packages.
Over time, these sites have diverged in how they go about doing this,
resulting in duplication of code, overly complex code, and incomplete
implementations. Installing a package involves several steps, and we
had some paths that did them all, but others that left out
some (important!) steps, like performing proper package validation.

This commit consolidates and simplifies all package installation logic
into a single place; everything flows through it, ensuring a
consistent experience.

Overall, when you install a package now, you provide either a package
identifier or the path to a local `.hart` file, an option that is now
formally represented in the type system (see below for more). Update
logic for "at-once" services is no longer coupled to the installation
process; all service updates are performed by the
`ServiceUpdater`. When you provide an identifier to the installation
functions, a call to Builder is made, and the latest appropriate
release is installed on the system. This always includes the download,
verification, and installation of all dependencies of the package.

In addition to the standard low-level fixes you'd expect (renaming
functions, fixing type signatures, adding docstrings, etc.), some of
the more significant or interesting changes include:

- Don't try to update a package when when installing from a spec
  We have an entire thread dedicated to updating services; we don't need
  two places that do this.

- `ServiceSpec`s no longer drive the installation process
  We use the common installation logic whether we've got a hart file or an
  identifier. We no longer install from a spec, but rather generate one
  after we install a package.

  This decouples `ServiceSpec`s, which are effectively run-time
  descriptors, from the installation process.

- Don't install when loading a spec in the manager

  There's no reason to go installing packages when we load a service
  spec in the manager. That package should have already been installed
  when the spec was added. If you go manually screwing around with
  Habitat's files under `/hab` then you're already causing problems for
  yourself (also, we don't necessarily carry this level of paranoia to
  every single operation we do).

  With the refactored, single code path for installing packages, we've
  removed the special-case logic for whether or not to update an
  "at-once" service (that should be the responsibility of the updater
  processes only). By not installing here, but just retrieving the
  installed package, we can eliminate extra unnecessary queries to the
  Builder service.

- A `Periodic` trait was extracted from `ServiceUpdater`
  We have several places in the codebase that perform periodic tasks
  on a schedule. This trait consolidates that logic in one place to
  facilitate reuse and make code easier to read.

- The `ServiceUpdater` and `SelfUpdater` processes both use the common
  installation logic.

  Previously, these two processes had their own incomplete
  re-implementation of installation. Now, they're just like everything
  else.

- Add InstallSource abstraction

  We can install from either a package
  identifier (e.g. "core/redis/3.2.4") or from a local `.hart`
  archive. Previously, this input was represented internally as a
  simple string type, which meant several places where we were trying
  to figure out which input that string really represented. Not only
  was this logic duplicated, but it also made downstream functions
  more complicated and confusing than they needed to be.

  Now, we capture this situation explicitly with our type system, named
  `InstallSource`.

- We return a `PackageInstall` when installing a package

  Previously we'd just return a `PackageIdent`, but in all cases, you
  really want a `PackageInstall` instead (it's in the name!). You can
  get the ident from that.

Signed-off-by: Christopher Maier <cmaier@chef.io>