[hab] Fix `binlink_all_in_pkg()` which did not work correct with rootfs.
This fixe a bug where the "binlink all" behavior would not work
correctly if a custom `rootfs` path was provided.

As part of the bugfix, a test for both the single and the all behavior
have been added. Yes, that looks like a lot of test code. However, this
is a first-step experiemtn with introducing more test code in the `hab`
subcommands where root permissions or other side effects don't get in
the way. We now have a slightly better avenue for testing with
`UI`-outputing code and even a way to test the contents of both output
streams.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>