Don't use coloring for UI messages
This is going to be non-interactive (i.e., log files), so we don't
need coloring.

(Previously, this module only used `UI.progress()`, which is a no-op
for non-TTY outputs. The code brought in for downloading new keys had
actual `UI`-based logging; we could either eliminate those statements
for now, or take the route in this commit.)

Signed-off-by: Christopher Maier <cmaier@chef.io>