Document how to create a new bootstrap bundle
Now more than one person can do it!

This also tweaks the creation script to always pull in the latest
stable `hab-launcher`. Previously, the version was hard-coded in the
script (because it changes relatively infrequently), but we can just
as the Depot for the latest stable version and achieve the same effect
while eliminating unnecessary script changes.

Signed-off-by: Christopher Maier <cmaier@chef.io>