[hab] Directly invoke updated docker exporter via exec.
This change performs the same install-on-use logic as `hab sup` and `hab
studio` subcommands. The underlying program, `hab-pkg-export-docker`
will be directly invoked and therefore:

```
hab pkg export docker --help
```

Will return the usage from `hab-pkg-export-docker` in the
`core/hab-pkg-export-docker` package

The `exec_subcommand_if_called()` function in the main module was
widened to look at 3 subcommands deep, rather the 2 subcommands prior
to this change. This was necessary as due to the subcommand depth of
`hab pkg export docker`. Future exporters should not need to do this.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>